Use SO_SNDBUF and SO_RCVBUF and 64bit file offset.

+++ src/file.h
@@ -30,6 +30,6 @@ typedef void (*FILE_WRITE_CALLBACK)(FILE_WRITTEN_DATA *data);
 ASYNC_FILE *file_new(COMPLETION_PORT *port, char *path);
 void file_close(ASYNC_FILE *file);
 
-void file_write(ASYNC_FILE *file, long position, BUFFER *buffer, int offset, int count, FILE_WRITE_CALLBACK callback, void *tag);
+void file_write(ASYNC_FILE *file, long long position, BUFFER *buffer, int offset, int count, FILE_WRITE_CALLBACK callback, void *tag);
 
 #endif
